requirements.txt
fastapi
uvicorn
groq
faiss-cpu
sentence-transformers
torch
transformers
numpy
python-dotenv
